# Conflicts:
#	ruby/package.json

+++ js/packages/eyes-storybook/test/fixtures/storybook-versions/mochaHooks.js
@@ -6,14 +6,14 @@ const path = require('path');
 
 exports.mochaHooks = {
   beforeAll: async () => {
-    if (!process.env.STORYBOOK_VERSION) {
-      process.env.STORYBOOK_VERSION = 'latest';
+    if (!process.env.APPLITOOLS_FRAMEWORK_VERSION) {
+      process.env.APPLITOOLS_FRAMEWORK_VERSION = 'latest';
     }
-    const storybookVersion = String(process.env.STORYBOOK_VERSION);
+    const storybookVersion = String(process.env.APPLITOOLS_FRAMEWORK_VERSION);
     const versionDir = path.resolve(__dirname, `./${storybookVersion}`);
     fs.readdirSync(versionDir).map(fileName => {
       const filePath = path.join(versionDir, fileName);
-      if (fs.existsSync(filePath) && fs.lstatSync(filePath).isDirectory()) {
+      if (fs.existsSync(filePath) && fs.lstatSync(filePath).isDirectory() && !filePath.includes('.storybook')) {
         fs.rmSync(filePath, { recursive: true })
       }
     });
@@ -25,7 +25,7 @@ exports.mochaHooks = {
     });
   },
   afterAll: async () => {
-    delete process.env.STORYBOOK_VERSION;
+    delete process.env.APPLITOOLS_FRAMEWORK_VERSION;
     delete process.env.INIT_CWD;
   }
 };
